requirements.txt
discord.py==2.3.2
python-dotenv==1.0.0
sqlalchemy==1.4.41
psycopg2-binary==2.9.9
aiohttp==3.8.5 